refactor: move ErrorTextBox to its own folder - preps for css modules migration
refactor: ErrorTextBox, Container, and ErrorIcon components
- moves ErrorTextBox and ErrorIcon to separate folders
- adds native Container component
- adds css modules for ErrorTextBox and Container
