+++ src/components/ui-base/Container/container.module.css
@@ -1,0 +1,5 @@
+.container {
+  max-width: 60ch;
+  margin-left: auto;
+  margin-right: auto;
+}
