+++ src/styles/variables.css
@@ -17,6 +17,11 @@
 
     /* semantic colors  */
     --amp-shadows-sm: 0 1px 2px 0 rgba(0, 0, 0, 0.05);
-    
     --amp-colors-border: var(--amp-colors-neutral-200);
+    --amp-colors-error-background: #FEF2F2;
+    --amp-colors-error-border: #FECACA;
+    --amp-colors-error-text: #991B1B;
+
+    /* misc */
+    --amp-default-border-radius: 4px;
 }
